transaction

+++ mysite03/src/main/java/mysite/service/GuestbookService.java
@@ -1,8 +1,10 @@
 package mysite.service;
 
+import mysite.repository.GuestbookLogRepository;
 import mysite.repository.GuestbookRepository;
 import mysite.vo.GuestbookVo;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -10,20 +12,36 @@ import java.util.List;
 public class GuestbookService {
 
     private GuestbookRepository guestbookRepository;
+    private GuestbookLogRepository guestbookLogRepository;
 
-    public GuestbookService(GuestbookRepository guestbookRepository) {
+    public GuestbookService(GuestbookRepository guestbookRepository, GuestbookLogRepository guestbookLogRepository) {
         this.guestbookRepository = guestbookRepository;
+        this.guestbookLogRepository = guestbookLogRepository;
     }
 
     public List<GuestbookVo> getContentsList() {
         return guestbookRepository.findAll();
     }
 
+    @Transactional
     public void deleteContents(Long id, String password) {
-        guestbookRepository.deleteByIdAndPassword(id, password);
+        GuestbookVo vo = guestbookRepository.findById(id);
+        if (vo == null) {
+            return;
+        }
+
+        int count = guestbookRepository.deleteByIdAndPassword(id, password);
+        if (count == 1) {
+            guestbookLogRepository.updateByRegDate(vo.getRegDate());
+        }
     }
 
+    @Transactional
     public void addContents(GuestbookVo vo) {
+        int count = guestbookLogRepository.update();
+        if (count == 0) {
+            guestbookLogRepository.insert();
+        }
         guestbookRepository.insert(vo);
     }
 }
